Add dmp methods, modify --config option.
1. For classes that already had an operator<< option to dump
it out, add a "dmp" method to dump to console, for use from
the debugger.

2. When the --config option is used, completely overwrite the
configuration. Previously the parameters part was backed up
and restored, but one use case for --config is when running
retdec-decompiler from the build directory in which case it
does not find the config json file, so there are no llvm
passes. So, if any options alter the configuration, they
should follow the --config option.